refactor: SearchTextFieldをcomponentsパッケージに移動
